requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests
python-dotenv==1.0.0
Pillow
pandas
supabase
langchain
langchain-community 
psycopg2
langchain-tavily  
langchain_cloudflare
python-dateutil
pycountry
fastapi
jamaibase>=0.3.0 
assemblyai>=0.28.0  